+++ src/main.rs
@@ -1,11 +1,13 @@
 use std::net::TcpListener;
+use crate::startup::run;
 
-use blogger::run;
+// use blogger::run;
+mod startup;
+mod routes;
 
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
-    let listener = TcpListener::bind("127.0.0.1:0")
-    .expect("Failed to bind to a random port");
+    let listener = TcpListener::bind("127.0.0.1:8000").expect(r#"Failed to bind."#);
 
     run(listener)?.await
 }
