+++ src/startup.rs
@@ -1,0 +1,16 @@
+use actix_web::{web, App, HttpServer, Responder, HttpResponse};
+use actix_web::dev::Server;
+use std::net::TcpListener;
+
+pub(crate) use crate::routes::{health_check::health_check, subscriptions::subscribe};
+
+pub fn run(listener: TcpListener) -> Result<Server, std::io::Error> {
+    println!("Executing run");
+    Ok(HttpServer::new(|| {
+        App::new()
+            .route("/health_check", web::get().to(health_check))
+            .route("/subscriptions", web::post().to(subscribe))
+    })
+    .listen(listener)?
+    .run())
+}
